feat: play a factor → play a part/role / be a factor
I also split the `phrase_set_corrections` into separate files for 1:1 linters and many:many linters as the test will was over 3,000 lines
