+++ harper-core/default_config.json
@@ -4050,7 +4050,7 @@
 							"Bool": {
 								"name": "AheadAnd",
 								"state": true,
-								"label": "Ahead And"
+								"label": "Ahead and"
 							}
 						},
 						{
@@ -5430,6 +5430,13 @@
 								"name": "PiqueInterest",
 								"state": true,
 								"label": "Pique Interest"
+							}
+						},
+						{
+							"Bool": {
+								"name": "PlayAFactor",
+								"state": true,
+								"label": "Play a Factor"
 							}
 						},
 						{
